refactor: move more components to the ui components directory
fix: fix import issues
