+++ src/components/Canary/index.tsx
@@ -17,6 +17,7 @@ import React, {
 import { FaFilter } from "react-icons/fa";
 import { useLocation, useSearchParams } from "react-router-dom";
 import { Badge } from "..";
+import HealthPageSkeletonLoader from "../../ui/SkeletonLoader/HealthPageSkeletonLoader";
 import { GroupByDropdown } from "../Dropdown/GroupByDropdown";
 import {
   DropdownStandaloneWrapper,
@@ -27,7 +28,6 @@ import { TimeRange, timeRanges } from "../Dropdown/TimeRange";
 import { defaultTabSelections } from "../Dropdown/lib/lists";
 import useRefreshRateFromLocalStorage from "../Hooks/useRefreshRateFromLocalStorage";
 import { isSidebarCollapsedAtom } from "../Layout/SidebarLayout";
-import HealthPageSkeletonLoader from "../SkeletonLoader/HealthPageSkeletonLoader";
 import { Toggle } from "../Toggle";
 import { TristateToggle } from "../TristateToggle";
 import { CanaryInterfaceMinimal } from "./CanaryInterface";
