+++ src/components/Configs/Sidebar/ConfigDetails.tsx
@@ -3,8 +3,8 @@ import { Link } from "react-router-dom";
 import { useGetConfigByIdQuery } from "../../../api/query-hooks";
 import { isCostsEmpty } from "../../../api/types/configs";
 import { Age } from "../../../ui/Age";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import { InfoMessage } from "../../InfoMessage";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { Status } from "../../Status";
 import DisplayDetailsRow from "../../Utils/DisplayDetailsRow";
 import { DisplayGroupedProperties } from "../../Utils/DisplayGroupedProperties";
