+++ src/components/Configs/Sidebar/ConfigsPanel.tsx
@@ -9,9 +9,9 @@ import { VscJson } from "react-icons/vsc";
 import { Tooltip } from "react-tooltip";
 import { useComponentConfigRelationshipQuery } from "../../../api/query-hooks/useComponentConfigRelationshipQuery";
 import { removeManualComponentConfigRelationship } from "../../../api/services/configs";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import EmptyState from "../../EmptyState";
 import { IconButton } from "../../IconButton";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
 import { toastError, toastSuccess } from "../../Toast/toast";
