+++ src/components/DataTable/index.tsx
@@ -20,8 +20,8 @@ import clsx from "clsx";
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { TiArrowSortedDown, TiArrowSortedUp } from "react-icons/ti";
 import usePreferences from "../../hooks/userPreferences";
+import TableSkeletonLoader from "../../ui/SkeletonLoader/TableSkeletonLoader";
 import { InfoMessage } from "../InfoMessage";
-import TableSkeletonLoader from "../SkeletonLoader/TableSkeletonLoader";
 import { DataTableRow } from "./DataTableRow";
 import { Pagination, PaginationType } from "./Pagination/Pagination";
