+++ src/components/Playbooks/Runs/PlaybookRunsSidePanel.tsx
@@ -9,10 +9,10 @@ import { useEffect, useMemo, useState } from "react";
 import { AiOutlineTeam } from "react-icons/ai";
 import { useNavigate } from "react-router-dom";
 import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import EmptyState from "../../EmptyState";
 import { PlaybookStatusIcon } from "../../Icon/PlaybookStatusIcon";
 import { InfiniteTable } from "../../InfiniteTable/InfiniteTable";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
 import PlaybookSpecIcon from "../Settings/PlaybookSpecIcon";
