+++ src/components/SchemaResourcePage/SchemaResource.tsx
@@ -1,12 +1,12 @@
 import { useNavigate, useParams } from "react-router-dom";
 import { useSettingsUpdateResource } from "../../api/query-hooks/mutations/useSettingsResourcesMutations";
 import { useGetSettingsResourceDetails } from "../../api/query-hooks/settingsResourcesHooks";
+import TableSkeletonLoader from "../../ui/SkeletonLoader/TableSkeletonLoader";
 import { BreadcrumbNav } from "../BreadcrumbNav";
 import EmptyState from "../EmptyState";
 import ErrorPage from "../Errors/ErrorPage";
 import { Head } from "../Head/Head";
 import { SearchLayout } from "../Layout";
-import TableSkeletonLoader from "../SkeletonLoader/TableSkeletonLoader";
 import { SchemaResourceEdit } from "./SchemaResourceEdit";
 import { SchemaResourceType } from "./resourceTypes";
