+++ src/components/Topology/Sidebar/ComponentChecks.tsx
@@ -5,9 +5,9 @@ import { useQuery } from "@tanstack/react-query";
 import { useAtom } from "jotai";
 import { useEffect } from "react";
 import { AiFillHeart } from "react-icons/ai";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import { CheckLink } from "../../Canary/HealthChecks/CheckLink";
 import EmptyState from "../../EmptyState";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
