+++ src/components/Topology/Sidebar/ComponentTeams.tsx
@@ -4,8 +4,8 @@ import { AiOutlineTeam } from "react-icons/ai";
 import { useGetComponentsTeamQuery } from "../../../api/query-hooks";
 import PillBadge from "../../../ui/Badge/PillBadge";
 import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import EmptyState from "../../EmptyState";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
 import { ComponentTeamLink } from "./ComponentTeamLink";
