+++ src/components/Topology/TopologyCard/index.tsx
@@ -5,13 +5,13 @@ import { Link, useParams, useSearchParams } from "react-router-dom";
 import { getTopology } from "../../../api/services/topology";
 import { Topology } from "../../../api/types/topology";
 import { Size } from "../../../types";
+import TopologyCardSkeletonLoader from "../../../ui/SkeletonLoader/TopologyCardSkeletonLoader";
 import AgentName from "../../Agents/AgentName";
 import { HealthChecksSummary } from "../../Canary/HealthChecksSummary";
 import { HealthSummary } from "../../Canary/HealthSummary";
 import { CustomScroll } from "../../CustomScroll";
 import { Icon } from "../../Icon";
 import IncidentCardSummary from "../../IncidentCardSummary";
-import TopologyCardSkeletonLoader from "../../SkeletonLoader/TopologyCardSkeletonLoader";
 import { CardMetrics } from "./CardMetrics";
 import { Property } from "./Property";
 import { TopologyConfigAnalysisLine } from "./TopologyConfigAnalysisLine";
