+++ src/components/Topology/TopologyConfigLinkModal/TopologyConfigLinkModal.tsx
@@ -12,10 +12,10 @@ import { Topology } from "../../../api/types/topology";
 import { queryClient } from "../../../query-client";
 import { Events, sendAnalyticEvent } from "../../../services/analytics";
 import { Modal } from "../../../ui/Modal";
+import TextSkeletonLoader from "../../../ui/SkeletonLoader/TextSkeletonLoader";
 import { delayedPromise, stringSortHelper } from "../../../utils/common";
 import ConfigLink from "../../Configs/ConfigLink/ConfigLink";
 import { DropdownWithActions } from "../../Dropdown/DropdownWithActions";
-import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import { toastError, toastSuccess } from "../../Toast/toast";
