+++ src/pages/TopologyPage.tsx
@@ -3,7 +3,6 @@ import { Topology } from "@flanksource-ui/api/types/topology";
 import { Head } from "@flanksource-ui/components/Head/Head";
 import { InfoMessage } from "@flanksource-ui/components/InfoMessage";
 import { SearchLayout } from "@flanksource-ui/components/Layout";
-import CardsSkeletonLoader from "@flanksource-ui/components/SkeletonLoader/CardsSkeletonLoader";
 import { refreshButtonClickedTrigger } from "@flanksource-ui/components/SlidingSideBar";
 import { toastError } from "@flanksource-ui/components/Toast/toast";
 import TopologySidebar from "@flanksource-ui/components/Topology/Sidebar/TopologySidebar";
@@ -11,6 +10,7 @@ import { TopologyBreadcrumbs } from "@flanksource-ui/components/Topology/Topolog
 import { TopologyCard } from "@flanksource-ui/components/Topology/TopologyCard";
 import TopologyFilterBar from "@flanksource-ui/components/Topology/TopologyPage/TopologyFilterBar";
 import { getCardWidth } from "@flanksource-ui/components/Topology/TopologyPopover/topologyPreference";
+import CardsSkeletonLoader from "@flanksource-ui/ui/SkeletonLoader/CardsSkeletonLoader";
 import { useQuery } from "@tanstack/react-query";
 import { useAtom } from "jotai";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
