+++ src/ui/SkeletonLoader/IncidentDetailsPageSkeletonLoader.tsx
