+++ src/ui/SkeletonLoader/TableSkeletonLoader.tsx
